requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
telethon==1.32.1
python-dotenv==1.0.0
nest-asyncio==1.5.8
greenlet==3.0.1
